id-9351 Less strict GPP local storage grant
The code for checking local storage grant was too strict checking for localStoragePurpose consent even when
there was no applicable section that needs this consent, or even in cases of sections we don't know how to handle.

Not we assume there is a LSP consent when there is no applicableSection which explicitly requires it.